backend/pyproject.toml
[project]
name = "rechnung-ng-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "pydantic[email]>=2.7",
    "pyyaml>=6.0",
    "filelock>=3.14",
    "weasyprint>=62",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
    "python-magic>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.4",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
